+++ cipher/Cargo.toml
@@ -13,12 +13,13 @@ categories = ["cryptography", "no-std"]
 
 [dependencies]
 generic-array = "0.14"
+crypto-common = { version = "0.1", path = "../crypto-common/" }
 
 blobby = { version = "0.3", optional = true }
 rand_core = { version = "0.6", optional = true }
 
 [features]
-std = []
+std = ["crypto-common/std"]
 dev = ["blobby"]
 
 [package.metadata.docs.rs]
